+++ CHANGELOG.md
@@ -20,11 +20,12 @@
 - Minify static assets in web
 - Replace snowfall with sakura flowers
 - Update Termux installation design
-- Whitelist internal requests in APILimiter
+- Whitelist requests in APILimiter to ignore internals
 - Start web when using `.weburl` and `--no-web` is used
 - Add inline commands targeted security rules control
 - Completely remove security groups `sudo` and `support`
 - Fix proxy-pass
+- New core modules: `UnitHeta`, `Translator`
 
 ## 🌑 Hikka 1.6.1